refactor: 토스트 전역 상태로 통합
