+++ src/App.tsx
@@ -5,11 +5,15 @@ import { RouterProvider } from 'react-router';
 import router from '@routes/router';
 
 import { queryClient } from '@apis/config/queryClient';
+
+import { ToastProvider } from '@components/toast/ToastProvider';
 import '@styles/global.css';
 
 const App = () => (
   <QueryClientProvider client={queryClient}>
-    <RouterProvider router={router} />
+    <ToastProvider>
+      <RouterProvider router={router} />
+    </ToastProvider>
     <ReactQueryDevtools initialIsOpen={false} />
   </QueryClientProvider>
 );
